Refactored InfoChartRender to manage chart traces and layout using useState and useEffect, ensuring that trace visibility is maintained when chart data updates.
